Update master branch server dev

+++ GameServer/AscensionProtocol/SubCode.cs
@@ -1,6 +1,9 @@
 ﻿namespace AscensionProtocol
 {
-    public enum SubCode:byte
+    /// <summary>
+    /// 二级操作码操作码
+    /// </summary>
+    public enum SubCode:short
     {
         GetRole,
         AddRole,
